+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Add a timeout for waiting on Unity threads (#515)
 * Add new SDL audio functions and update SDL_dynapi (#518)
 * RAM search can be interrupted
+* Hook mono sleep function
 
 ### Changed